services/agent-economy/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
asyncpg>=0.29.0
httpx>=0.27.0
pydantic>=2.7.0
PyJWT>=2.8.0
